examples/08_financial_analyst/requirements.txt
crewai
crewai-tools
mcp
yfinance
pandas
pyyaml
python-dotenv
langchain-core
